# lib/api-zod/src/generated/types/message.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for AI Doubt Solver & Notes Generator
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const Message = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.enum(["user", "assistant"]),
  content: zod.string(),
  createdAt: zod.string(),
});

export type Message = zod.input<typeof Message>;