# lib/api-zod/src/generated/types/pdfUploadResult.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for AI Doubt Solver & Notes Generator
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const PdfUploadResult = zod.object({
  text: zod.string(),
  filename: zod.string(),
  pageCount: zod.number(),
  wordCount: zod.number(),
});

export type PdfUploadResult = zod.input<typeof PdfUploadResult>;